requirements.txt

streamlit
openai-whisper
torch
